fix: UploadChatAttachments and remove map
